[client] add new_community_call function
